Added a command to figure out which could be the next voice of seren
According to the wiki https://runescape.wiki/w/Voice_of_Seren#Randomness_of_the_voice
it will never be the current voice of seren, or the previous one. Leaving us with 4 options.
So basically we look at the history and just return all clans but the current voice of seren
and the previous one.