Added a start on runemetrics profile parsing
The plan is to be able to parse most of the common events, for now
I only implemented a simple parser for boss kills. As of now I have
only tested this with some bosses from my own recent kills. On which
it seemed to work fine. I see no reason why it wouldn't work with all
bosses, but edge cases be edge cases..